Test heroku

+++ src/main/java/com/tomc/sysapp/model/parkingFares/impl/ParkingFare.java
@@ -1,7 +1,5 @@
 package com.tomc.sysapp.model.parkingFares.impl;
 
-import java.math.BigDecimal;
-
 import com.tomc.sysapp.model.parkingFares.ParkingFareTypeEnum;
 import com.tomc.sysapp.model.vehicle.CarTypeEnum;
 
@@ -9,9 +7,9 @@ public class ParkingFare {
 
 	private ParkingFareTypeEnum fareType;
 	private CarTypeEnum carType;
-	private BigDecimal fare;
+	private long fare;
 
-	public ParkingFare(ParkingFareTypeEnum fareType, CarTypeEnum carType, BigDecimal fare) {
+	public ParkingFare(ParkingFareTypeEnum fareType, CarTypeEnum carType, long fare) {
 		this.fareType = fareType;
 		this.carType = carType;
 		this.fare = fare;
